+++ docs/changes.md
@@ -1,5 +1,7 @@
 # 最新动态
 2022/01/06
+  * fscript增加random0to1。
+  * fscript支持until/repeat_times/repeat语句。
   * fscript中增加is\_int/is\_odd/is\_even等函数
   * 还原awtk\_main.inc文件的循环代码（感谢智明提供补丁）